liquid: convert order book to TS
